Countries manager form completed

+++ MahjongTournamentSuite/MahjongTournamentSuite/Ranking/RankingForm.cs
@@ -13,7 +13,7 @@ namespace MahjongTournamentSuite.Ranking
         private const int WM_NCLBUTTONDOWN = 0xA1;
         private const int HT_CAPTION = 0x2;
         [DllImportAttribute("user32.dll")]
-        private static extern long SendMessage(IntPtr hWnd, long Msg, long wParam, long lParam);
+        private static extern int SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
         [DllImportAttribute("user32.dll")]
         private static extern bool ReleaseCapture();
