+++ MahjongTournamentSuite/MahjongTournamentTimer/FormTimer.cs
@@ -12,7 +12,7 @@ namespace MahjongTournamentTimer
         private const int HT_CAPTION = 0x2;
 
         [DllImportAttribute("user32.dll")]
-        private static extern long SendMessage(IntPtr hWnd, long Msg, long wParam, long lParam);
+        private static extern int SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
         [DllImportAttribute("user32.dll")]
         private static extern bool ReleaseCapture();
